+++ packages/for-ui/src/menu/MenuItem.tsx
@@ -14,7 +14,7 @@ export const MenuItem = forwardRef<HTMLLIElement, MenuItemProps>((props, ref) =>
       ref={ref}
       classes={{
         root: fsx([
-          'bg-shade-white-default text-r text-shade-dark-default hover:bg-shade-white-hover whitespace-nowrap border-solid py-2 pl-6 pr-12 font-sans',
+          `bg-shade-white-default text-r text-shade-dark-default hover:bg-shade-white-hover whitespace-nowrap border-solid py-2 pl-6 pr-12 font-sans focus-visible:bg-shade-white-active [&.Mui-focused]:bg-shade-white-active`,
           className,
         ]),
       }}