+++ packages/for-ui/src/menu/MenuList.tsx
@@ -1,14 +1,8 @@
 import { forwardRef } from 'react';
 import MuiMenuList, { MenuListProps } from '@mui/material/MenuList';
 import { fsx } from '../system/fsx';
+import { style } from './style';
 
 export const MenuList = forwardRef<HTMLUListElement, MenuListProps>(({ className, ...rest }, ref) => (
-  <MuiMenuList
-    ref={ref}
-    className={fsx(
-      `z-modal shadow-menu w-full rounded p-1 divide-shade-light-default grid grid-cols-1 divide-y`,
-      className
-    )}
-    {...rest}
-  />
+  <MuiMenuList ref={ref} className={fsx(style, className)} {...rest} />
 ));