+++ packages/for-ui/src/select/Select.stories.tsx
@@ -30,7 +30,7 @@ const options: SelectOption[] = [
   },
 ];
 
-const SKILL = {
+const SKILL: Record<string, string> = {
   // C#
   csharp: 'csharp',
   // C++
@@ -96,7 +96,7 @@ export const Basic: Story = () => {
                 label="国名"
                 placeholder="国名"
                 options={options}
-                onChange={(e, option) => {
+                onChange={(_, option) => {
                   onChange((option as SelectOption)?.inputValue);
                 }}
               />
@@ -115,18 +115,17 @@ export const Basic: Story = () => {
           control={control}
           render={({ field: { name, onChange } }) => {
             return (
-              <>
-                <Select
-                  name={name}
-                  required
-                  placeholder="国名"
-                  options={options}
-                  onChange={(e, option) => {
-                    onChange((option as SelectOption)?.inputValue);
-                  }}
-                />
-                <FormHelperText error>入力してください</FormHelperText>
-              </>
+              <Select
+                name={name}
+                required
+                placeholder="国名"
+                options={options}
+                onChange={(_, option) => {
+                  onChange((option as SelectOption)?.inputValue);
+                }}
+                error
+                helperText="入力してください"
+              />
             );
           }}
         />
@@ -160,7 +159,7 @@ export const Single: Story = () => {
               label="国名"
               placeholder="未定"
               options={options}
-              onChange={(e, option) => {
+              onChange={(_, option) => {
                 onChange(option as SelectOption);
               }}
             />
@@ -196,7 +195,7 @@ export const Disabled: Story = () => {
               label="国名"
               placeholder="国名"
               options={options}
-              onChange={(e, option) => {
+              onChange={(_, option) => {
                 onChange((option as SelectOption)?.inputValue);
               }}
             />
@@ -258,9 +257,7 @@ export const Multiple: Story = () => {
                   label="国名"
                   placeholder="未定"
                   options={options}
-                  onChange={(e, option) => {
-                    onChange((option as SelectOption)?.inputValue);
-                  }}
+                  onChange={onChange}
                 />
                 <FormHelperText error>入力してください</FormHelperText>
               </>
@@ -309,7 +306,7 @@ export const MultipleFreeSolo: Story = () => {
               name={name}
               value={value}
               options={skillOptions}
-              onChange={(e, option) => {
+              onChange={(_, option) => {
                 onChange(option);
                 // onChange((option as SelectOption)?.inputValue)
               }}
@@ -347,7 +344,7 @@ export const DisableFilter: Story = () => {
               label="国名"
               placeholder="未定"
               options={options}
-              onChange={(e, option) => {
+              onChange={(_, option) => {
                 onChange(option as SelectOption);
               }}
             />